Service integration testing for centos7 (systemd)
Docker vs SystemD!

Setup a Container as VM for centos7 platform:
Add a specific Dockerfile for centos7 that will install systemd, sshd
and create kitchen user. See the original source in this blog post for details:
https://medium.com/brigade-engineering/reduce-chef-infrastructure-integration-test-times-by-75-with-test-kitchen-and-docker-bf638ab95a0a#.f9yglinhz

Add a monkeypatch for `Kitchen::Driver::Docker#generate_keys` to only
generate keys if private and public keys are non-nil. This allow us to
disable publickey authentication method and use the default password.
Maybe this will be changed by a fixed ssh key

Fixed test-kitchen to 1.4.2